backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pydantic[email-validator]==2.10.3
pydantic-settings==2.7.0
gitpython==3.1.43
kubernetes==31.0.0
pyyaml==6.0.2
httpx==0.28.1
